+++ Imperium/src/Interface/MapUI/MinimapSettings.cs
@@ -11,15 +11,23 @@ internal class MinimapSettings : BaseUI
 {
     protected override void InitUI()
     {
-        ImpToggle.Bind("Content/Gizmos/ShowInfoPanel", transform, Imperium.Settings.Map.MinimapInfoPanel, theme);
-        ImpToggle.Bind("Content/Gizmos/ShowLocationPanel", transform, Imperium.Settings.Map.MinimapLocationPanel, theme);
+        ImpToggle.Bind("Content/Gizmos/ShowInfoPanel", container, Imperium.Settings.Map.MinimapInfoPanel, theme);
+        ImpToggle.Bind("Content/Gizmos/ShowLocationPanel", container, Imperium.Settings.Map.MinimapLocationPanel, theme);
 
         ImpSlider.Bind(
             path: "Content/Scale",
-            container: transform,
+            container: container,
             valueBinding: Imperium.Settings.Map.MinimapScale,
             indicatorFormatter: value => $"{value:0.0}",
             theme: theme
+        );
+
+        ImpButton.Bind(
+            "TitleBox/Close",
+            container,
+            onClick: Close,
+            theme: theme,
+            isIconButton: true
         );
     }
 
@@ -27,9 +35,17 @@ internal class MinimapSettings : BaseUI
     {
         ImpThemeManager.Style(
             themeUpdate,
-            transform,
+            container,
+            new StyleOverride("TitleBox", Variant.DARKER),
+            new StyleOverride("Border", Variant.DARKER),
             new StyleOverride("Content/Width/Overlay", Variant.LIGHTER),
             new StyleOverride("Content/Height/Overlay", Variant.LIGHTER)
+        );
+
+        ImpThemeManager.StyleText(
+            themeUpdate,
+            container,
+            new StyleOverride("TitleBox/Title", Variant.FOREGROUND)
         );
     }
 }